Placed css in package styles
