+++ STONKS/screens/Add.js
@@ -1,20 +1,13 @@
 import React, {useEffect, useState} from "react";
-import {
-    StyleSheet,
-    Text,
-    View,
-    TextInput,
-    TouchableOpacity,
-    Keyboard,
-    TouchableWithoutFeedback,
-} from "react-native";
+import {Text, View, TextInput, TouchableOpacity, Keyboard, TouchableWithoutFeedback,} from "react-native";
 import getCurrentDate from "../utils/getDate.js";
 import Scan from './Scan.js';
 import API_URL from "../url";
 import List from "../components/List";
 import DateTimePickerModal from "react-native-modal-datetime-picker";
 import ItemListProduct from "../components/ItemListProduct";
-import addElement from "../request/addElement"
+import addElement from "../request/addElement";
+import styles from "../styles/AddStyle";
 
 
 export default function Add({route, navigation}) {
@@ -145,51 +138,6 @@ export default function Add({route, navigation}) {
     );
 }
 
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        backgroundColor: "#fff",
-        alignItems: "center",
-        justifyContent: "flex-start",
-    },
-    text: {
-        marginTop: '8%',
-        fontSize: 30,
-        fontWeight: 'bold',
-    },
-    bAdd: {
-        marginTop: '8%',
-        backgroundColor: 'darkseagreen',
-        padding: 10,
-        borderRadius: 5,
-        width: '50%',
-        height: '15%',
-        alignItems: 'center',
-        justifyContent: 'center',
-    },
-    scanText: {
-        color: '#fff',
-        fontSize: 25,
-        fontWeight: 'bold',
-        alignItems: 'center',
-        justifyContent: 'center',
-    },
-    dropDown: {
-        backgroundColor: '#0e131e',
-        padding: 20,
-        paddingLeft: 15,
-        paddingRight: 15,
-        fontSize: 22,
-    },
-    selection: {
-        padding: 20,
-        textAlign: 'center',
-        fontSize: 28,
-    },
-    datePickerStyle: {
-        width: 200,
-        marginTop: 20,
-    },
-});
+
