+++ STONKS/screens/DisplayInfo.js
@@ -1,15 +1,16 @@
-import {StyleSheet, Text, TouchableOpacity, View} from 'react-native';
+import {Text, TouchableOpacity, View} from 'react-native';
 import React, {useEffect, useState} from "react";
 import API_URL from "../url";
 import moveElement from "../request/moveElement";
 import getLocation from "../request/getLocation";
+import styles from "../styles/DisplayInfoStyle.js";
 
 export default function DisplayInfo({navigation, route}) {
-
+    // get data
     const {element, location, token} = route.params;
-
+    // state to handle location change
     const [isLocationChange, setLocationChange] = useState();
-
+    // state to store location name
     const [locationName, setLocationName] = useState(element.location.name)
 
     const changeLocation = () => {
@@ -83,63 +84,3 @@ export default function DisplayInfo({navigation, route}) {
 }
 
 
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        backgroundColor: "white",
-        alignItems: "center",
-        justifyContent: "flex-start",
-    },
-    iView: {
-        flex: 3,
-        marginTop: '10%',
-        backgroundColor: "white",
-        alignItems: 'center',
-        flexDirection: 'column',
-    },
-    text: {
-        fontSize: 25,
-        marginBottom: '8%',
-        fontWeight: 'bold',
-        alignItems: 'center',
-        justifyContent: 'center',
-    },
-    textL:{
-        fontSize: 25,
-        marginBottom: '10%',
-        fontWeight: 'bold',
-        alignItems: 'center',
-        justifyContent: 'center',
-    },
-    bView: {
-        flex: 1.8,
-        alignItems: 'center',
-        flexDirection: 'column',
-        justifyContent: 'space-between',
-        marginBottom: '15%',
-    },
-    bText: {
-        fontSize: 25,
-        color: '#fff',
-        textAlign: 'center',
-        fontWeight: 'bold',
-    },
-    button: {
-        justifyContent: 'center',
-        alignItems: 'center',
-        backgroundColor: 'lightskyblue',
-        padding: 15,
-        borderRadius: 5,
-        width: '100%',
-        height: '40%',
-    },
-    bLocation: {
-        justifyContent: 'center',
-        alignItems: 'center',
-        backgroundColor: 'darkseagreen',
-        padding: 15,
-        borderRadius: 5,
-        width : '100%',
-        height: '40%',
-    }
-});
