+++ STONKS/styles/DisplayInfoStyle.js
@@ -1,0 +1,64 @@
+import {StyleSheet} from "react-native";
+
+const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        backgroundColor: "white",
+        alignItems: "center",
+        justifyContent: "flex-start",
+    },
+    iView: {
+        flex: 3,
+        marginTop: '10%',
+        backgroundColor: "white",
+        alignItems: 'center',
+        flexDirection: 'column',
+    },
+    text: {
+        fontSize: 25,
+        marginBottom: '8%',
+        fontWeight: 'bold',
+        alignItems: 'center',
+        justifyContent: 'center',
+    },
+    textL:{
+        fontSize: 25,
+        marginBottom: '10%',
+        fontWeight: 'bold',
+        alignItems: 'center',
+        justifyContent: 'center',
+    },
+    bView: {
+        flex: 1.8,
+        alignItems: 'center',
+        flexDirection: 'column',
+        justifyContent: 'space-between',
+        marginBottom: '15%',
+    },
+    bText: {
+        fontSize: 25,
+        color: '#fff',
+        textAlign: 'center',
+        fontWeight: 'bold',
+    },
+    button: {
+        justifyContent: 'center',
+        alignItems: 'center',
+        backgroundColor: 'lightskyblue',
+        padding: 15,
+        borderRadius: 5,
+        width: '100%',
+        height: '40%',
+    },
+    bLocation: {
+        justifyContent: 'center',
+        alignItems: 'center',
+        backgroundColor: 'darkseagreen',
+        padding: 15,
+        borderRadius: 5,
+        width : '100%',
+        height: '40%',
+    }
+});
+
+export default styles;
