+++ STONKS/styles/ScanStyle.js
@@ -1,0 +1,25 @@
+import {StyleSheet} from "react-native";
+
+const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        flexDirection: 'column',
+        justifyContent: 'center',
+        height: '100%',
+        width: '100%',
+    },
+    text: {
+        fontSize: 35,
+        fontWeight: 'bold',
+    },
+    button: {
+        padding: 20,
+        textAlign: 'center',
+        fontSize: 28,
+        alignItems: 'center',
+        justifyContent: 'center',
+    },
+
+});
+
+export default styles;
